More restructuring for readability
