+++ slacksocketmode/deadman.go
@@ -1,0 +1,7 @@
+package slacksocketmode
+
+import "time"
+
+func deadmanDuration(d time.Duration) time.Duration {
+	return d * 4
+}
