+++ slacksocketmode/slacksocketmode_test.go
